Reorganize RegisterView code into multiple more specialized files
