+++ plugins/ODbgRegisterView/ODbgRV_x86Common.h
@@ -1,0 +1,27 @@
+/*
+Copyright (C) 2015 Ruslan Kabatsayev <b7.10110111@gmail.com>
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 2 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <http://www.gnu.org/licenses/>.
+*/
+
+#ifndef X86_COMMON_H_20170817
+#define X86_COMMON_H_20170817
+
+constexpr unsigned FPU_TAG_EMPTY = 3;
+
+static constexpr auto FSR_NAME = "FSR";
+static constexpr auto FCR_NAME = "FCR";
+static constexpr auto FTR_NAME = "FTR";
+
+#endif
